scripts/requirements.txt
psycopg2-binary==2.9.10
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.0
